requirements.txt
bcrypt==3.2.2
Brotli==1.0.9
cffi==1.15.1
click==8.1.3
contextlib2==21.6.0
cryptography==37.0.4
Flask==2.1.2
Flask-Compress==1.12
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
PyJwt==2.4.0
MarkupSafe==2.1.1
pycparser==2.21
PyMySQL==1.0.2
schema==0.7.5
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2